+++ html/js/bypass.js
@@ -16,3 +16,9 @@ var aliasMeasurement = {
 "hour": "AVG last hour",
 "day": "AVG last 24h"
 };
+
+// spacename aliases, need to be inserted manually once
+var aliasSpacenames = {
+    "livlab": "Office 1F",
+    "office": "Office 2F"
+};